+++ cli/cli/Commands/Organization/RealmListCommand.cs
@@ -103,11 +103,39 @@ public class RealmListCommand : AtomicCommand<RealmsListCommandArgs, RealmsListC
 	{
 		try
 		{
-			var api = args.Provider.GetService<IRealmsApi>();
-			var customer = await api.GetCustomer();
-			var game = await api.GetGame(args.AppContext.Pid);
-			var orgRealmData = game.projects.ToRealmArray(args.AppContext.Cid, args.IncludeArchived);
-			
+			var customer = await args.Provider.GetService<IRealmsApi>().GetAdminCustomer();
+			var orgRealmData = customer.customer.projects
+				.Where(r => args.IncludeArchived || !r.archived)
+				.Select(r =>
+			{
+				var rd = new OrgRealmData();
+				rd.Cid = args.AppContext.Cid;
+				rd.Pid = r.name;
+				rd.RealmName = r.displayName;
+				rd.RealmSecret = r.secret;
+				rd.ParentPid = r.parent.GetOrElse("");
+
+				var parentRealm = r.parent.GetOrElse("");
+				rd.IsProduction = string.IsNullOrEmpty(parentRealm);
+				var projectName = rd.IsProduction ? r.displayName.GetOrElse("") : "";
+
+				// We count how many parents this realm has --- 1 means its the staging realm; two means it's a dev realm.
+				var parentCount = 0;
+				while (!string.IsNullOrEmpty(parentRealm))
+				{
+					parentCount += 1;
+					var p = customer.customer.projects.First(p => p.name == parentRealm);
+					parentRealm = p.parent.GetOrElse("");
+
+					projectName = string.IsNullOrEmpty(parentRealm) ? p.displayName : "";
+				}
+
+				rd.IsStaging = parentCount == 1;
+				rd.IsDev = parentCount > 1;
+				rd.ProjectName = projectName;
+				return rd;
+			}).ToArray();
+
 			// Get the token and fill it out args.AppContext.Token.
 			return new RealmsListCommandOutput() { CustomerAlias = customer.customer.alias.GetOrElse(""), Cid = customer.customer.cid.ToString(), VisibleRealms = orgRealmData, };
 		}