+++ cli/cli/Services/UnrealSourceGenerator/UnrealApiSubsystemDeclaration.cs
@@ -208,7 +208,7 @@ class ₢{nameof(UnrealSourceGenerator.exportMacro)}₢ UBeam₢{nameof(Subsyste
 {{
 private:
 	GENERATED_BODY()
-	/** @brief Initializes the auto-increment Id and binds the ExecuteRequestDelegate to DefaultExecuteRequestImpl  */
+	/** @brief Initializes the auto-increment Id */
 	virtual void Initialize(FSubsystemCollectionBase& Collection) override;
 
 	/** Cleans up the system.  */